+++ src/Buckle/Buckle/CodeAnalysis/Binding/BinderFlags.cs
@@ -10,4 +10,5 @@ internal enum BinderFlags : uint {
     None,
     LocalFunction = 1 << 0,
     LowLevelRegion = 1 << 1,
+    Method = 1 << 2,
 }
